+++ src/interfaces/IFlashAggregator.sol
@@ -5,7 +5,7 @@ interface IFlashAggregator {
     function flashLoan(
         address[] memory _tokens,
         uint256[] memory _amounts,
-        uint256 _route,
+        uint16 _route,
         bytes calldata _data,
         bytes calldata
     ) external;
